Task#10 - refactoring structure dao and service, added getting model from dto in service
