+++ src/main/java/com/itechart/javalab/library/controller/command/ajax/impl/LendBookCommand.java
@@ -5,8 +5,8 @@ import com.itechart.javalab.library.controller.command.ajax.AjaxCommand;
 import com.itechart.javalab.library.controller.util.json.impl.JacksonJsonBuilder;
 import com.itechart.javalab.library.controller.util.json.JsonBuilder;
 import com.itechart.javalab.library.dto.BorrowRecordDto;
-import com.itechart.javalab.library.service.ReaderService;
-import com.itechart.javalab.library.service.impl.DefaultReaderService;
+import com.itechart.javalab.library.service.BorrowRecordService;
+import com.itechart.javalab.library.service.impl.DefaultBorrowRecordService;
 import lombok.extern.log4j.Log4j2;
 
 import javax.servlet.ServletException;
@@ -19,14 +19,14 @@ import static com.itechart.javalab.library.controller.util.ResponseParameterName
 @Log4j2
 public class LendBookCommand implements AjaxCommand {
 
-    private final ReaderService readerService;
+    private final BorrowRecordService readerService;
     private final JsonBuilder jsonBuilder;
     private static final String REQUEST_PARAMETER_EDITED_RECORDS = "addedRecords";
     private static final String RESPONSE_MESSAGE_PARTLY_FAILED = "partlyFailed";
     private static final String RESPONSE_MESSAGE_OK = "ok";
 
     public LendBookCommand() {
-        this.readerService = DefaultReaderService.getInstance();
+        this.readerService = DefaultBorrowRecordService.getInstance();
         this.jsonBuilder= JacksonJsonBuilder.getInstance();
     }
