+++ src/main/java/com/itechart/javalab/library/controller/command/front/impl/GetBookCommand.java
@@ -6,9 +6,9 @@ import com.itechart.javalab.library.dto.BookPageDto;
 import com.itechart.javalab.library.model.Book;
 import com.itechart.javalab.library.model.BorrowRecord;
 import com.itechart.javalab.library.service.BookService;
-import com.itechart.javalab.library.service.ReaderService;
+import com.itechart.javalab.library.service.BorrowRecordService;
 import com.itechart.javalab.library.service.impl.DefaultBookService;
-import com.itechart.javalab.library.service.impl.DefaultReaderService;
+import com.itechart.javalab.library.service.impl.DefaultBorrowRecordService;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -20,13 +20,13 @@ import java.util.Optional;
 public class GetBookCommand implements Command {
 
     private final BookService bookService;
-    private final ReaderService readerService;
+    private final BorrowRecordService readerService;
 
     private static final String BOOK_PAGE_DTO = "bookPageDto";
 
     public GetBookCommand() {
         this.bookService = DefaultBookService.getInstance();
-        this.readerService = DefaultReaderService.getInstance();
+        this.readerService = DefaultBorrowRecordService.getInstance();
     }
 
     @Override
