+++ src/main/java/com/itechart/javalab/library/dao/ReaderDao.java
@@ -1,25 +1,16 @@
 package com.itechart.javalab.library.dao;
 
-import com.itechart.javalab.library.model.BorrowRecord;
+import com.itechart.javalab.library.model.Paginator;
 import com.itechart.javalab.library.model.Reader;
 
-import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
 public interface ReaderDao {
 
-    Optional<List<BorrowRecord>> getBorrowRecords(int bookId);
+    Optional<Set<Reader>> getReaders(Paginator paginator);
 
-    boolean setBorrowRecordStatus(List<BorrowRecord> borrowRecord);
-
-    boolean createBorrowRecord(List<BorrowRecord> borrowRecords);
-
-    boolean updateStatusBorrowRecords(List<BorrowRecord> borrowRecords);
+    Optional<Integer> getNumberReadersRecords();
 
     Optional<Set<Reader>> getReadersByEmail(String email);
-
-    List<BorrowRecord> getReturnNotificationInfo();
-
-    List<BorrowRecord> getDelayNotificationInfo();
 }
