--- src/main/java/com/itechart/javalab/library/dao/ReceiveReaderDao.java
@@ -1,14 +1,0 @@
-package com.itechart.javalab.library.dao;
-
-import com.itechart.javalab.library.model.Paginator;
-import com.itechart.javalab.library.model.Reader;
-
-import java.util.Optional;
-import java.util.Set;
-
-public interface ReceiveReaderDao {
-
-    Optional<Set<Reader>> getReaders(Paginator paginator);
-
-    Optional<Integer> getNumberReadersRecords();
-}
